chore: Refactor base code
