+++ backend/crm/src/test/java/cn/cordys/crm/system/controller/ProductControllerTests.java
@@ -71,9 +71,6 @@ class ProductControllerTests extends BaseTest {
     void testModuleField() throws Exception {
         MvcResult mvcResult = this.requestGetWithOkAndReturn(MODULE_FORM);
         moduleFormConfig = getResultData(mvcResult, ModuleFormConfigDTO.class);
-
-        // 校验权限
-        requestGetPermissionTest(PermissionConstants.PRODUCT_MANAGEMENT_READ, MODULE_FORM);
     }
 
     @Test
